Move loadModeledMethods to a separate file
