+++ extensions/ql-vscode/src/data-extensions-editor/data-extensions-editor-view.ts
@@ -20,8 +20,6 @@ import {
   showAndLogExceptionWithTelemetry,
   showAndLogErrorMessage,
 } from "../common/logging";
-import { readFile } from "fs-extra";
-import { load as loadYaml } from "js-yaml";
 import { DatabaseItem, DatabaseManager } from "../databases/local-databases";
 import { CodeQLCliServer } from "../codeql-cli/cli";
 import { asError, assertNever, getErrorMessage } from "../common/helpers-pure";
@@ -33,7 +31,6 @@ import { showResolvableLocation } from "../databases/local-databases/locations";
 import { decodeBqrsToExternalApiUsages } from "./bqrs";
 import { redactableError } from "../common/errors";
 import { readQueryResults, runQuery } from "./external-api-usage-query";
-import { loadDataExtensionYaml } from "./yaml";
 import { ExternalApiUsage } from "./external-api-usage";
 import { ModeledMethod } from "./modeled-method";
 import { ExtensionPack } from "./shared/extension-pack";
@@ -44,9 +41,8 @@ import {
 } from "./auto-model";
 import { enableFrameworkMode, showLlmGeneration } from "../config";
 import { getAutoModelUsages } from "./auto-model-usages-query";
-import { getOnDiskWorkspaceFolders } from "../common/vscode/workspace-folders";
 import { Mode } from "./shared/mode";
-import { saveModeledMethods } from "./modeled-method-fs";
+import { loadModeledMethods, saveModeledMethods } from "./modeled-method-fs";
 
 export class DataExtensionsEditorView extends AbstractWebview<
   ToDataExtensionsEditorMessage,
@@ -189,45 +185,14 @@ export class DataExtensionsEditorView extends AbstractWebview<
 
   protected async loadExistingModeledMethods(): Promise<void> {
     try {
-      const extensions = await this.cliServer.resolveExtensions(
-        this.extensionPack.path,
-        getOnDiskWorkspaceFolders(),
+      const modeledMethods = await loadModeledMethods(
+        this.extensionPack,
+        this.cliServer,
+        this.app.logger,
       );
-
-      const modelFiles = new Set<string>();
-
-      if (this.extensionPack.path in extensions.data) {
-        for (const extension of extensions.data[this.extensionPack.path]) {
-          modelFiles.add(extension.file);
-        }
-      }
-
-      const existingModeledMethods: Record<string, ModeledMethod> = {};
-
-      for (const modelFile of modelFiles) {
-        const yaml = await readFile(modelFile, "utf8");
-
-        const data = loadYaml(yaml, {
-          filename: modelFile,
-        });
-
-        const modeledMethods = loadDataExtensionYaml(data);
-        if (!modeledMethods) {
-          void showAndLogErrorMessage(
-            this.app.logger,
-            `Failed to parse data extension YAML ${modelFile}.`,
-          );
-          continue;
-        }
-
-        for (const [key, value] of Object.entries(modeledMethods)) {
-          existingModeledMethods[key] = value;
-        }
-      }
-
       await this.postMessage({
         t: "loadModeledMethods",
-        modeledMethods: existingModeledMethods,
+        modeledMethods,
       });
     } catch (e: unknown) {
       void showAndLogErrorMessage(
